requirements.txt
scikit-learn>=1.0.0
requests>=2.25.0
tqdm>=4.60.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.12.0
adjustText>=0.8
wordcloud>=1.8.1
python-dateutil>=2.8.0
packaging>=21.0
upsetplot>=0.6.0 